Re-structure and clean up the code, also add documentation.
