--- Octo-Bot/file-transfer-bot/ftp-bot/README
@@ -1,4 +1,0 @@
-# Octo-Bot: FTP bot
-
-## FTP Upload & Download
-
