--- Octo-Bot/file-transfer-bot/ftp-bot/bot/README.txt
@@ -1,1 +1,0 @@
-FTP Bot
